create html boilerplate, stylesheet & add images
flexbox.html
linked styles.css to the page, this page will be used to practice with
flexbox. A menu container has been created with a menu with items
inside: date, signup, login. This will be the first iteration for this
tutorial project.

styles.css
a universal style has been created where the margin and padding will be set to 0 while the box-sizing will be set to border-box. added
colors to the menu container as well with a padding on the top and
bottom. the items for the menu will be contained with a white border
given a width of 900px.

images/
added a folder with images, they will be added in future updates.