Create header container, apply styles to new .menu-container
flexbox.html
A new header-container has been created with 3 items: subscribe & 2
images.

styles.css
The header-container and header have been styled similar to what is
used for the menu-container and its items. The container is styled
to have its items to be centered horizontally as well with a height
of 300px