Spaced menu container more optimally
flexbox.html
Deleted the links div to space out the signup and login in a more
optimal way using flexbox properties instead.

styles.css
All that was needed in this change was to use the margin-left property
set to auto. This will give out the max space between date and signup.